refactor: move the non-interactive scan phase out of the program config
Mirror self-driving's layout: the CI/headless-specific detection
machinery lives in posthog-integration/detect-agentic.ts (sibling of
detect.ts, same shape as self-driving/detect-agentic.ts) and index.ts
just imports and calls it — the program config's diff is one import and
one line in ciPreRun.

Generated-By: PostHog Code
Task-Id: 06aaaf67-5aba-497e-bd9e-0220e2417442
